chore: migrate to f32

+++ src/constant.rs
@@ -1,5 +1,5 @@
 pub const VECTOR_DIMENSION: usize = 2;
-pub const PLOT_GRAPH_X_Y_RANGE: (std::ops::Range<f64>, std::ops::Range<f64>) =
+pub const PLOT_GRAPH_X_Y_RANGE: (std::ops::Range<f32>, std::ops::Range<f32>) =
     (0.0..2000.0, 0.0..2000.0);
-pub const VECTOR_VALUE_RANGE: std::ops::Range<f64> = 0.0..2000.0;
+pub const VECTOR_VALUE_RANGE: std::ops::Range<f32> = 0.0..2000.0;
 pub const SEED_DATASET_SIZE: u32 = 500;
